Introduce `TextInputActions` for TextInput
Summary:
This diff is to address the issue from our support group(https://fb.workplace.com/groups/litho.support/permalink/3493344677653542/), where reports that we're missing some APIs for the Kotlin version of `TextInput`.
Given that we have a bunch of text event handlers, I believe it is easier to scale and manage them by grouping them into a class.

Reviewed By: zielinskimz

Differential Revision: D44702898

fbshipit-source-id: acbb5acfad995bf1774e1810c4cb6ca32edec334