Add ability to use a different check logic for the cancellation inside runAndGet
Summary:
While trying to experiment the resolve cancellation with interruption of futures, we have been seeing some weird MIDs that I can't reproduce or reason about.

I'm adding a configuration that defines how we check if an exception should be attributed to a forceful cancellation of the future. I want to experiment with this to verify if it affects the appearance of these weird MIDs.

Reviewed By: zielinskimz

Differential Revision: D44831693

fbshipit-source-id: 63ce0ab34f20a730887de833e040c8a55d44f142