Preload Primitive classes on startup
Summary:
In order to test the perf impact of Primitive API we need the classes to be present in betamap or preloaded at startup. I've tried to run an experiment with a beta gk so that these classes are picked up by betamap but it was unsuccesful.
This diff manually preloads Primitive classes at app startup. It'll be removed once finish performance testing of Primitive API.

Reviewed By: adityasharat

Differential Revision: D44950550

fbshipit-source-id: 246751a41a2246ef910d4487aa90bf3f1ca9032d