Introduce mDebugData on render tree, Store most recent layout result in debug builds
Summary:
For UIDebugger => Bloks support we need to be able to get access to the most recent LayoutResult, For litho there is Layout state but this is currently not stored anywhere for bloks.

The most natural place to store the layout result is the render tree. This approach works for both the RootHostView and RenderTreeHost view as they both have a render tree.

There was already a mutable property called mRenderTreeData of type object. It was used for setting the litho layout state but it was never accessed. This property has been renamed to mDebugData to make its intention clearer. For bloks / lite this is set to the layout result in debug builds only to not increase memory pressure.

The litho state was never read out of the property so it was removed. It can be readded later if needed

Reviewed By: adityasharat

Differential Revision: D45048024

fbshipit-source-id: 59ac219c31d21c3f643456492d4c42270f74cad3