Fix the incorrect logic of using layout cache
Summary:
This diff is to fix a bug while using layout cache, we should compare the size specs across different measure pass.

We also don't need to record the size info for component that has exact size, because If a component has exactly the same size spec for different layout processes, we won't be able to reach layout phase.

Reviewed By: adityasharat

Differential Revision: D45527856

fbshipit-source-id: bfae20a414fd0565ec958db0f5a6d1d8dcc84aaa