Force remount in experimental variant
Summary:
We identified a bug in the *LithoUsageMode.Experimental* where `Drawables` are not being re-draw after a view isRecycled and then re-attached. This is due to an inconsistency in the contract of binding/mount whenever we avoid unmounting the items.

To mend this, we are now marking the `MountState` as needing to remount whenever we recycle a view.

Reviewed By: astreet

Differential Revision: D45568994

fbshipit-source-id: a50f10919cac2125b4e8ff9e27d9bb338827de6f