pre-fill RV pool with feed Litho views
Summary:
## Context
The Feed Foundation team is working to convert a few high frequency rows in Instagram Feed, to Litho. One current blocker is that the existing feed, has optimizations to pre-fill the RV Pool with Views that we know will be rendered on feed.

We can execute this same optimization with Litho, by pre-filling the pool with Image and Text specs that we know the UFI and Media Topic header use.

Differential Revision: D45569229

fbshipit-source-id: 925f9ae1ecdb46e720555d23d23714a2ad428dd7