Remove outdated code
Summary:
# Context
I did a bit search to understand this piece of code and noticed that it was introduced (D3586160 / D3593199) to collect logs for a special case in mid-2016, and then logging was removed(D7813600) in 2018 because it's not used at all. I think we're safe to remove this code for  simplicity.

Reviewed By: adityasharat

Differential Revision: D45984299

fbshipit-source-id: 66907d1e71178b104926b45174d50d354024ca68