Use BindData API to store onUnbind lambdas
Summary:
This diff uses new BindData feature to fix Primitive API bug where Binders were storing UnbindFunc instances. This caused issues when component was re-rendered. Thanks to BindData API, Binders can be pure funcions and the UnbindFunc is now stored in MountItem and passed to Binder.unbind when necessary.

Added a test that illustrates the issue, prior to this change that test was failing because the onUnbind{} lambdas weren't invoked.

Reviewed By: astreet, colriot

Differential Revision: D46015985

fbshipit-source-id: a26b19fa7d73e360359fdc8358903a1d5c6a2285