Refactor RenderUnit.Binder interface
Summary:
This diff refactors `RenderCore.Binder` interface, so that the bind method returns a an object which represents bind data, and the unbind method accepts that bindData object created during bind.

All callsites have been updated accordingly.

The logic that stores and retrieves this bind data will be added in subsequent diff. This one is only about the codemod.

[RFC](https://docs.google.com/document/d/1aBO5jtv6YNpB5Kdx3Zc81JUmJAeIGBzh0Y7G51q09sc/edit#heading=h.tybb42xfx2qf)

Reviewed By: astreet

Differential Revision: D45863209

fbshipit-source-id: 06eb6f5cae9076652bc885c2619eb58078b1fa94