Add ability to integrate a Events listener into a ComponentTree
Summary:
As part of the process of replacing `ComponentsLogger` with an API that relies on the new Events API, I'm adding the ability to listen to events associated with a ComponentTree. These events can be ComponentTree/RenderCore related and they will use the `renderStateId/componentTreeId` as the reference.

## Client Side

The client will only define a `ComponentTreeDebugEventListener` where it specifies the action to perform on an `DebugEvent` and the set of events it is interested in.

## Internals

Whenever a listener is present, we will wrap it inside a `DebugEventSubscriber` which will only propagate the events related to that ComponentTree and for which the listener is interested in. This subscriber will be automatically subscriber and will unsubscribe during `release`.

Reviewed By: adityasharat

Differential Revision: D46080849

fbshipit-source-id: 2c2ade11b89feede22e037615aca915a3d3a1288