Fix the bug that MountSpec with fixed size doesn't get rebinding
Summary:
## Context
After we introduced reusing render unit, there's a case that has been missed. Since all components that have fixed size will not be measured by Yoga, and we only recreate render unit inside the `measure` call, which is fine to Mountable and Primitive because we will invoke `measure` explicitly for them later, but that's not the case for MountSpec.

## Fix
Recreate render unit for unmeasured MountSpec as well in `measurePendingSubtrees`.

Reviewed By: adityasharat

Differential Revision: D46590010

fbshipit-source-id: 882c718272c0f81321afd2e958a4eb4dfbde3fc5