Add ComponentTree Id information to CalculationStateContext
Summary:
So that we can provide the `componentTreeId` in `DebugEvent` related to *Resolve/Prepare* we need to make that information available in the `CalculationStateContext`.

There was a tricky situation to solve though related which led me to need to add the `treeId` information to the `LithoTree` (but below I leave a comment on that specific piece of code - in case there are other suggestions).

Reviewed By: adityasharat

Differential Revision: D46385909

fbshipit-source-id: 4ac97d1a7d46ac2730b1d0db3cfcd8d1edfacef8