Remove classes from litho plugin that is not compatible with AS flamingo
Summary:
Right now the plugin fails to build against AI-222(flamingo) and the errors comes from the OnEventChangeSignatureDialog.java. As Kata mentioned in D45840328, it is fine to remove this class as the team is currently rewriting this .

Also removed OnEventGenerateAction.java and OnEventCreateFix.java since they are referencing OnEventChangeSignatureDialog.

Reviewed By: Katalune

Differential Revision: D46675811

fbshipit-source-id: f2dcfd60edfccbabd4910499c50d3c6e6db4b446