Only unmounts Litho components in disappearing animations
Summary:
Only unmounts Litho components in disappearing animations.
This diff fixes a bug where the transition extension was unmounting items from a nested root host (i.e. NOT managed by the target mount state; specifically a bloks hierarchy in this reference use case). This was happening because of an incorrect check to ignore unsupported hosts.

This diff fixes the issue by ignore any RenderUnit that is not a TransitionRenderUnit, and any host that is not a DisappearingHost.

Reviewed By: nicous

Differential Revision: D46801629

fbshipit-source-id: 10a0351eb32dc39a5450ba904ead3c33fc33421f