Fix bindDynamic API by using the DV as a binder dep
Summary: This diff fixes the `bindDynamic` API by using the dynamic value instance as a dependency in the underlying binder. Prior to that change, when we had a dynamic value and there was a state update (even unrelated to the DV), then the DV was always rebound. Now the DV will be unbound and bound again only if the DV instance has changed between the state updates.

Reviewed By: adityasharat

Differential Revision: D47019696

fbshipit-source-id: 1b04cf4d605b71e45b06719a0f39908490103d21