Remove `TextInputActions` to simplify `TextInput`
Summary:
# Context
We introduced `TextInputActions` in D44702898 to add missing APIs for `TextInput`, but which may not be the best way to expose these APIs because it ended up having to include a container class. We could leverage Kotlin's Named Arguments to achieve the same goal without the need to introduce one more container.

Reviewed By: colriot

Differential Revision: D46858555

fbshipit-source-id: a550bfbfa4e17c9986816275ebb5fb8ff9efabf4