Make bindDynamic accept nullable values
Summary:
As per title - prior to this change, when the DynamicValue was nullable, then there was no way to use it because `bindDynamic` APIs accepted non-nullable type and we can't use conditional checks because we have linters that disallow code like `dynamicValue?.let { bindDynamic(dynamicValue, ...) }`.

This change fixes the API by making `bindDynamic` accept nullable types.

Reviewed By: pentiumao, adityasharat

Differential Revision: D47020627

fbshipit-source-id: 9a6c5fc314fa294999f417696997beedc96c6de9