Move UI Debuggers to global module
Summary:
The goal of this diff is to move the debuggers created to support Litho IG Migration into a general purpose `litho-tooling` module so we can re-use across apps and integrate with different features if needed.

Any other future UI Debugger can eventually be implemented here.

P.s ideally we will have most the current functionality in the official UI Debugger, but this is our fallback until that is done.

Reviewed By: zielinskimz

Differential Revision: D46931109

fbshipit-source-id: 93cd5ac26741c016db5f7125abbd4387ee231267