Fix the issue that `OnCreateTreeProp` invoked twice when reconciliation is enabled
Summary:
# Context
We have a bug that `OnCreateTreeProp` will be invoked twice when reconciliation is enabled. By looking into this issue, I noticed that root cause is like this:

1. `getReconciliationMode` relies on a param of `ComponentContext`, which comes from `reconcile`.
2. In order to meet the signature, we're preparing a new `ComponentContext` before calling `reconcile`.
3. I think that's the only reason for we create scoped context here and never use it for other purposes.

https://www.internalfb.com/code/fbsource/[406794ab4801f472c89a79577d0eb54a05b0d10d]/fbandroid/libraries/components/litho-core/src/main/java/com/facebook/litho/Resolver.java?lines=93-94

Based on the fact that the nullable check inside `getReconciliationMode` is meaningless because Kotlin can ensure that for us, we can get rid of this param.

https://www.internalfb.com/code/fbsource/[406794ab4801f472c89a79577d0eb54a05b0d10d]/fbandroid/libraries/components/litho-core/src/main/java/com/facebook/litho/ScopedComponentInfo.kt?lines=26

Then we can avoid `createScopedContext` before `reconcile`, and `OnCreateTreeProp` would never be invoked twice any more.

Reviewed By: adityasharat

Differential Revision: D46855360

fbshipit-source-id: 6abc68c405f75738b0d9f0b8a04b6010fba73acc