Adds infra-structure to better debug tracking issues
Summary:
This diff focuses in enabling and building some logging that allows the identification of situations where some Litho content is being mounted and there is some tracking data for that content present in the *LogContextRegistry*.

This situation is not problematic if it's just the same content being mounted (e.g, the icon replacing another icon, because the tracking data will be updated). However, there were some situations where a `ComponentHost` is being mounted and the previously tracking data was not cleared up, and won't be replaced at all. These situations will lead to bugs and we should have easy ways to identify them.

In a next diff I will integrate this debugger behind a developer option so that it is easily pluggable.

Reviewed By: adityasharat

Differential Revision: D47212628

fbshipit-source-id: 83fd363c0fc28ebfcd44b6277c87900877dec6f3