Revert D47197949: Multisect successfully blamed "D47197949: Move mCommonProps to SpecGeneratedComponent" for test or build failures
Summary:
This diff is reverting D47197949
D47197949: Move mCommonProps to SpecGeneratedComponent by mkarpio has been identified to be causing the following test or build failures:

Tests affected:
- [//fbandroid/instrumentation_tests/com/facebook/workchat/threadlist/thread/components:components - build](https://www.internalfb.com/intern/test/281475072333001/)

Here's the Multisect link:
https://www.internalfb.com/multisect/2395943
Here are the tasks that are relevant to this breakage:

We're generating a revert to back out the changes in this diff, please note the backout may land if someone accepts it.

If you believe this diff has been generated in error you may Commandeer and Abandon it.

Reviewed By: mkarpio

Differential Revision: D47301149

fbshipit-source-id: 256fb2a2fa58da6470185288ba55ab173d0023aa