Add support to a custom mount binder
Summary:
# Context
There are two sets of problems with ads tracking in Litho surfaces that were being exposed with the migration for Feed:

- Whenever *Component Host Recycling* is enabled, we were seeing that the tracking data was not correctly being cleared up.
- If more tracked "Components" would be shared between surfaces (e.g. Reels and Feed) the problem above would also appear.

For this reason, we are looking to provide a reliable way to track and untrack view data. Another important point, is that the current tracking framework uses `View` as keys. Therefore, we need to get access to the `View` to which each `Component` materializes into.

# Solution
We are going to allow custom `RenderUnit.Binder` that hook into the Litho mount phase. This way, we will have "callbacks" that can be called everytime some content is mounted and unmounted.

Therefore, we need to provide the ability to setup a *Mount* `RenderUnit.Binder`. This diff focuses on the Litho part of that integration.

It adds an extension to `Style`, which allows us to specify the `RenderUnit.Binder`:

```
Style.viewBinder(binder) or Style.drawableBinder(binder)
```

For the IG situations, we would end up using `viewBinder` because their logging framework uses `View` as keys. Therefore, if we have some kind of `Drawable` being mounted we need to guarantee that we correctly get a hosting `View` to use in the logging framework.

Reviewed By: adityasharat

Differential Revision: D47322639

fbshipit-source-id: aa6c30e2b387e723d574c082bf5b3f1c1ccc0dca