added fallback toString when no editor
Summary:
Added a fallback when an editor cant be found. Bit of a bandaid but its very simple and tt serves 2 purpose:
1. the user knows there are props there since we never skip serialization due to no editor found
2. if they are using a data class we get a reasonable serialization thanks to kotlin. In debug builds the toSTring isNot stripped out

In future id like to look into a proper data serializer for a proper data class serializer for a better experience in flipper but for now this will have to do

As a result of 1. we will need to keep an eye on the payload size and serialization time metrics. at least this all happens on a background thread.

Reviewed By: fabiocarballo

Differential Revision: D47296200

fbshipit-source-id: d4aad0f16a85c0fb91a8fcfccf1f82328732568f