Revert Use RenderUnit.getExtra API to store info about excluding from IM and doesMountRenderTreeHosts
Summary:
This diff reverts changes from D46723014 and D46834627 because of: T158054484 and T158020506.

I'm not removing `RenderUnit.getExtra` API because it can be useful for other things as well. I've removed APIs that depended on `RenderUnit.getExtra` and fixed all callsites.

Will try to add this logic back, but behind a QE.

Reviewed By: adityasharat

Differential Revision: D47583912

fbshipit-source-id: ca925bc5fa5c309e3db7c5779b5efdd5e07b4377