Make it possible to disable pooling by setting pool size to 0
Summary: As per title - this diff changes the pooling logic, so that when pool size is set to 0 on `ContentAllocator`, then the pool is not created at all and thus pooling is disabled and content is created every time. Prior to that change, passing 0 pool size resulted in a crash in Pool constructor.

Reviewed By: adityasharat

Differential Revision: D47549106

fbshipit-source-id: 2bb8d386edf4aa4204086e51d80b5c615c28599d