Simplify Some Assertion Methods with Streams
Summary:
There's a couple things happening here, but in short, it uses streams in a couple places (which is a bit closer to the library functions you'd use if/when this file is converted to Kotlin).

In a lot more words, I used a stream to replace the for loop for containsAll that way we'd be able to just call the Java class version.  THEN, I realized we had a function that did that already, so I just replaced it's implementation with the stream one.

Okay, great, and then I just updated some of the other ones to use streams as well.

We don't have to use streams if people prefer for loops and if statements though.

Reviewed By: adityasharat

Differential Revision: D47621415

fbshipit-source-id: 82644df036ef0c9e4081ee57173be0c08740e31f