Improve ContainsComponent Error Message to Avoid Empty List
Summary:
I actually want to make this even better (and print the actual view tree for the Litho View), but we don't support lazy messages in our current `AssertJ` version.

Therefore, this at least improves the message like we did for `DirectComponent` that way you don't have the less useful:
```
Expecting:
[]
to have exactly 1 times <Component with type <class com.facebook.litho.Row>>
```
which, understandably, is a bit confusing (and not useful).

 ---

Reviewed By: mkarpio

Differential Revision: D47892448

fbshipit-source-id: dc16f0524757655905686df101bdba720d5a723a