Make LazyList reconciliation default consistent with LithoConfiguration
Summary:
The signature of the `LazyList` is defaulting the `reconciliationEnabled` to `false`. This means that it will overlook the definition in the `LithoConfiguration` and that it will be disabled for any `Component` inside such hierarchy.

I'm changing the default to be the same value as defined in the `LithoConfiguration`. (This change is gated just in case there is any problem and we have to roll it back immediately).

Reviewed By: adityasharat

Differential Revision: D48066201

fbshipit-source-id: b37639e865b3caec458af9fe8980e12f722140e4