Avoid creating content when pre-allocating if pool is already full
Summary:
I have noticed that in the `MountItemsPool.prefillMountContentPool` we pass a `poolSize` as a number of items to pre-allocate. However,
it is not guaranteed that the given `MountContent` actually has an `ItemPool` of that given size.

Therefore, it can happen that `MountA` has a `poolSize=5` and we try to call this API with `poolSize=10`. This will lead to the redundant creation
of 5 mount contents that will not actually be allocated.

Therefore, I have added a check to stop the pre-allocation process early if we notice that the item was not successfully released into the pool. Ideally, this could be further optimized by checking before hand the pool size, but this is not available in the interface.

I will explore this hypothesis after.

Reviewed By: adityasharat

Differential Revision: D48112428

fbshipit-source-id: 22de0d00e8176a496033e90ed2144f31234272aa