Fix live editting
Summary:
This feature got damaged in a few refactors, the first issue was that we were bailing out if there was no time machine, even if the request update was not a time machine roll back

The second issue was that we try to always apply state override and if there was no state container we would crash.

I also noticed we dont apply the layout overrides anymore, they appear to not be neeed as i am able to adjust layout parameters and they take effect. Not sure if there is more to this though

Reviewed By: adityasharat

Differential Revision: D48036334

fbshipit-source-id: 32d7ff6d44666bf5b3ba9653de4ebad060a42e3f