Revert D48228032: Multisect successfully blamed "D48312650: [litho][layout-caching] Optimises caching to improve reuse of render units" for test or build failures
Summary:
This diff is reverting D48228032
D48312650: [litho][layout-caching] Optimises caching to improve reuse of render units by adityasharat has been identified to be causing the following test or build failures:

Tests affected:
- [//fbandroid/instrumentation_tests/com/facebook/messaging/business/customerfeedback/view:view - fiveStars_withoutFollowUp_lightScheme (com.facebook.messaging.business.customerfeedback.view.CustomerFeedbackRootComponentUITest)](https://www.internalfb.com/intern/test/562949973472957/)

Here's the Multisect link:
https://www.internalfb.com/multisect/2779589
Here are the tasks that are relevant to this breakage:

We're generating a revert to back out the changes in this diff, please note the backout may land if someone accepts it.

If you believe this diff has been generated in error you may Commandeer and Abandon it.

Differential Revision: D48319276

fbshipit-source-id: ac52196a64e14987618aa14a653ae141c033ee51