Use DefaultItemPool always as the default pool on Litho Components
Summary:
We were having a bug where `Primitives` wouldn't be pre-allocated because they defaulted to the `MountItemsPool.DefaultItemPool`, which had a no-op for `maybePreallocateContent`. This didn't happen for `MountSpecs` because we actually delegate to `DefaultMountContentPool` in Litho.

In order to solve this issue, I'm making the `MountItemsPool.DefaultItemPool` to contain the same logic that previously lived in the `DefaultMountContentPool` and making it the reference to be used in both `MountSpecs/Primitives` or any client that needs an ItemPool.

Reviewed By: astreet

Differential Revision: D48189905

fbshipit-source-id: d866fcabe9aa9249423ff7410fdc5363285a22eb