Add logs to help identify the root cause of incorrect measure spec
Summary:
## Context

We're facing a UBN(T160770211) task where the stack shows that a nested Recycler is throwing an exception of incorrect measure spec. However, we couldn't figure out where the use case is until print the view hierarchy out.
This diff is to add more logs to print the view hierarchy, helping us to identify where the problem is.

Reviewed By: fabiocarballo

Differential Revision: D48392117

fbshipit-source-id: 5218e5543ba43cd08cc1050a0f180072216c0edd