Add Quantity String Extension Function
Summary:
Adds a getter for the quantity string, applying it to both instances of `ResourceResolver`.

Separately though, while I was here, it seems like we could improve the resource cache to actually cache more strings?  Right now, we don't do it when there are format args (or quantities), but if we actually built the format args the `Key`, we could actually cache a bit more.  (Generally speaking, format args are almost always strings, but if it's something without an equals comparison, then it would mean we'd never match -- like some spannable strings.)

Not sure if there's historical context to not exploring that type of keying, but it *definitely* is easier to do in Kotlin than Java.

---------
## Internal:

Workplace post: https://fb.workplace.com/groups/litho.support/posts/3589155674739108

Reviewed By: fabiocarballo, zielinskimz

Differential Revision: D48458404

fbshipit-source-id: e02cff619ca93a4123d0df568635aab79b5de8b8