truncate text to constrained measures
Summary:
Android text grows unbounded unless it is given a strict number of lines to truncate to. This has the undesirable effect of text getting cut off when measurement constraints are bounded (`EXACT` & `AT_MOST`). On iOS, this is not the case and text is automatically constrained to the provided measurement bounds provided that no max lines constraints have been set.

This diff attempts to replicate the iOS behavior on Android.

It is important to note that we need another layout step to truncate and fit the text to constrained bounds. This means that the number of text layout calls is at least 2 and at most 3 (if you have background color props) in a single `calculateLayout` pass for text components. As layout is not cheap especially for large text, this could hurt performance on surfaces. For this reason, I'm putting this behavior behind a flag to allow for incremental adoption and possibly large scale experimentation.

Reviewed By: rooju

Differential Revision: D48341988

fbshipit-source-id: be33d1202c0c8a7fc49b1e9c0d3768627c550c0b