Ship converting SmartCropReboundViewPager to Primitive
Summary:
This change was added behind a MC to ensure that it doesn't cause any regressions. [The results](https://fburl.com/deltoid3/2dfkyukx) are neutral, so we're shipping this change.

This diff migrates `SmartCropReboundViewPager` from Litho Mountable API to Litho Primitive API. Mountable API is being replaced by the new Primitive API. There should be no functional or visual changes, it's the same component implemented with a new API.

Reviewed By: fabiocarballo

Differential Revision: D48384408

fbshipit-source-id: c0041dfda468d82e10a766744d15d6e20541b8b5