Standardize how arcs are drawn
Summary:
Standardise how we draw arcs across Android and iOS implementations.

The position of start/end angles on the circle should not change based on whether the value of the “clockwise” prop is true or false.

0 degrees is at the 3 o’clock position, and positive start/end degrees should be measured clockwise relative to the 0 degree position, while negative start/end degrees should be measured counterclockwise.

Right now a start/end angle of “x” is interpreted as “-x” if clockwise = false, which is wrong.

The value of the clockwise prop should only influence the direction of the sweep from the start to the end angle positions.  iOS does this correctly and so does HTML5.

 {F1075706262}

The iOS fixes related to arcs is here: D48608905

Reviewed By: zielinskimz

Differential Revision: D48601137

fbshipit-source-id: ccbf0d349bf0b25252d77126e73d3c545be0d7f9