Add ability to experiment between sync and non sync mount item pool
Summary:
After we have unified the implementation of `MountItemPool` between rendercore and litho, we are left with one single difference. Litho used synchronized pools, while rendercore didn't. For the sake of simplicity, we are going to test using non-synchronized pools in Litho as well.

This behavior is only needed for *MountSpecs* as Primitives/Mountables already rely on the render core default.

Reviewed By: zielinskimz

Differential Revision: D48731650

fbshipit-source-id: f8e7635aaa4e16cc585e5ce8db6d47c58c003b8e