Use LruCache for storing cached values
Summary:
As per title - this diff uses Android's `LruCache` for storing cached values computed using `useCached` hook. Prior to this change, the cached values were stored in an unbounded `HashMap` which means that once created, these values were kept in cache as long as component tree exists.
Also there was a single cache shared across all components in the tree.

In order to reduce the memory footprint, we're replacing `HashMap` with `LruCache`. We also create separate `LruCache` per `useCached` hook in a given component. The key in the `LruCache` are the deps passed to `useCached(deps){}` method.

In this diff the old logic is still in place in order to allow for experimentation and picking a good default cache size. Once we prove that this performs well, we'll remove the old `HashMap` and we'll make it possible for the developers to override the default cache size.

Reviewed By: adityasharat

Differential Revision: D49010732

fbshipit-source-id: 5298d8ac144fea7b60beb1cef8c5732fb6aff6b3