Fix inverse transforms & make them consistent
Summary:
**Background**
The existing type for defining inverse transforms is `bk.types.CanvasTransformInverse`
and it takes a `transform` prop.

However, iOS and Android interpret this very differently. Consider the following example:

```
$awesome_transform =
<bk:types:canvas-transform>
  <bk:types:canvas-transform-translate ... />
  <bk:types:canvas-transform-scale .../>
   ...
  <bk:types:canvas-transform-inverse transform={$some_transform}/>
</bk:canvas-transform>
```

The children inside the `bk:types:canvas-transform` are meant to be applied in order (post-concatenated, in matrix terms).

On Android, when the client applies the inverse transform element, it discards the matrix from prior transforms and instead just inverts the transform specified in the `transform` prop and applies that.

On iOS, the `transform` attribute is completely ignored and it inverts the accumulated result of the transforms that are its peers.

Since these are fundamentally different interpretations, zielinskimz, ivanmisuno & I discussed  how we'd like to define inverse transforms and came to the following solution:

1. Children elements of `bk:types:canvas-transform` should continue to be applied in order.

2. If a `bk:types:canvas-transform-inverse` is encountered, it inverts the accumulated transforms.

This means, we don't need the `transform` attribute on the `bk:types:canvas-transform-invert` element.

The fixes span both www and Android.  The iOS client already does this correctly, so no changes are needed.

## WWW (D49054439)
* Introduce new spec type: `bk.types.CanvasTransformInverseV2`
* Make `bk.types.CanvasTransformInverse` impl-only: `bk:types:canvas-transform-inverse-v1:impl`
* Introduce derived XHP data class: `bk:types:canvas-transform-inverse` which reduces to `bk:types:canvas-transform-inverse-v1:impl` on iOS and `bk:types:canvas-transforms-inverse-v2:impl` on Android

## Android (this diff)
* Bridges `bk.types.CanvasTransformInverseV2`
* Unbridges the buggy `bk.types.CanvasTransformInverse`, since the server won't send it anymore (see www section above)

## iOS
* No changes needed. The implementation of `bk.types.CanvasTransformInverse` does what we want already.

Reviewed By: zielinskimz

Differential Revision: D49058946

fbshipit-source-id: 0204611ddc0d93d553619236dfe15b9c022aa6d4